requirements.txt
boto3>=1.34.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
openpyxl>=3.1.0
